+++ src/hooks/useInfiniteScroll.js
@@ -1,0 +1,21 @@
+import { useEffect, useRef } from "react";
+
+export function useInfiniteScroll(fetchNextPage, hasNextPage) {
+  const observerRef = useRef();
+
+  useEffect(() => {
+    const observer = new IntersectionObserver(
+      (entries) => {
+        if (entries[0].isIntersecting && hasNextPage) {
+          fetchNextPage();
+        }
+      },
+      { threshold: 1 },
+    );
+
+    if (observerRef.current) observer.observe(observerRef.current);
+    return () => observer.disconnect();
+  }, [fetchNextPage, hasNextPage]);
+
+  return observerRef;
+}
