+++ src/utils/queryUtils.js
@@ -1,0 +1,13 @@
+export function updateSearchParams(
+  setSearchParams,
+  sortBy,
+  sortOrder,
+  searchQuery,
+) {
+  const params = {};
+  if (sortBy) params.sort = sortBy;
+  if (sortOrder) params.sortOrder = sortOrder;
+  if (searchQuery) params.q = searchQuery;
+
+  setSearchParams(params);
+}
